added proecss var and post route not working
