+++ lab-roger/lib/server.js
@@ -15,6 +15,8 @@ const MONGODB_URI = process.env.MONGODB_URI;
 // Middleware
 app.use(cors());
 app.use('/api/v1', router);
+require('../route/route-photo')(router);
+require('../route/route-gallery')(router);
 require('../route/route-auth')(router);
 app.all('/{0,}', (req, res) => errorHandler(new Error('Path Error. Route not found.'), res));
