+++ lab-roger/package.json
@@ -39,13 +39,18 @@
   "author": "",
   "license": "MIT",
   "dependencies": {
+    "aws-sdk": "^2.190.0",
     "bcrypt": "^1.0.3",
     "body-parser": "^1.18.2",
     "cors": "^2.8.4",
+    "crypto": "^1.0.1",
+    "del": "^3.0.0",
     "dotenv": "^5.0.0",
     "express": "^4.16.2",
     "jsonwebtoken": "^8.1.1",
     "mongoose": "^5.0.3",
+    "multer": "^1.3.0",
+    "path": "^0.12.7",
     "superagent": "^3.8.2"
   },
   "devDependencies": {
